pyproject.toml
[project]
name = "clew"
version = "0.1.0"
description = "Semantic code search with hybrid retrieval and MCP integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ruminaider", email = "hello@ruminaider.com"}
]
keywords = ["clew", "embeddings", "qdrant", "mcp", "claude"]

dependencies = [
    # Core
    "qdrant-client>=1.9.0,<2.0",
    "voyageai>=0.3.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",

    # Tokenization
    "transformers>=4.40.0",
    "tokenizers>=0.19.0",

    # AST Parsing
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-javascript>=0.21.0",

    # Config & Validation
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",

    # Ignore patterns
    "pathspec>=0.12.0",

    # Async & Retry
    "tenacity>=8.2.0",
    "httpx>=0.27.0",

    # MCP
    "mcp>=1.0.0",
]

[project.optional-dependencies]
nl = [
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "respx>=0.21.0",
    "anthropic>=0.40.0",
]

[project.scripts]
clew = "clew.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: requires external services (Qdrant)"]

[tool.mypy]
python_version = "3.10"
strict = true
